pos-backend-django/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
gunicorn
whitenoise
dj-database-url
